+++ ttt.hpp
@@ -1,4 +1,4 @@
-int checkGameOverP1(short result1, short result2, short result3, short result4, short result5, short result6, short result7, short result8, short result9);
-int checkGameOverP2(short result1, short result2, short result3, short result4, short result5, short result6, short result7, short result8, short result9);
+short checkGameOverP1(short result1, short result2, short result3, short result4, short result5, short result6, short result7, short result8, short result9);
+short checkGameOverP2(short result1, short result2, short result3, short result4, short result5, short result6, short result7, short result8, short result9);
 void printRules(void);
 void printCurrentBoard(short result1, short result2, short result3, short result4, short result5, short result6, short result7, short result8, short result9);
